+++ tests/constructor.rs
@@ -68,3 +68,38 @@ mod deprecated {
         field: i32,
     }
 }
+
+mod proxy_lint_attr {
+    #![deny(non_snake_case, clippy::too_many_arguments)]
+
+    use super::*;
+
+    #[allow(clippy::too_many_arguments)]
+    #[derive(Constructor)]
+    struct ManyArguments(
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+        u8,
+    );
+
+    #[expect(non_snake_case)]
+    #[derive(Constructor)]
+    struct User {
+        Num: i32,
+    }
+}